Moved time related classes to a time package
